+++ backend/substrapp/tasks/tasks_compute_task.py
@@ -171,8 +171,8 @@ def _run(
     # In case of retries: only execute the compute task if it is not in a final state
     with get_orchestrator_client(channel_name) as client:
         task = client.query_task(task.key)
-        # Set allow_doing=True to allow celery retries.
-        task_utils.abort_task_if_not_runnable(task.key, client, allow_doing=True, task=task)
+        # Set allow_executing=True to allow celery retries.
+        task_utils.abort_task_if_not_runnable(task.key, client, allow_executing=True, task=task)
         # Try to set the tasks status to DOING if it is not already the case
         task_utils.start_task_if_not_started(task, client)
 
@@ -201,8 +201,8 @@ def _run(
         with acquire_compute_plan_lock(compute_plan_key):
             # Check the task/cp status again, as the task/cp may not be in a runnable state anymore
             with get_orchestrator_client(channel_name) as client:
-                # Set allow_doing=True to allow celery retries.
-                task_utils.abort_task_if_not_runnable(task.key, client, allow_doing=True)
+                # Set allow_executing=True to allow celery retries.
+                task_utils.abort_task_if_not_runnable(task.key, client, allow_executing=True)
 
             # start inputs loading timer
             timer.start()